==> process/4_analyze.qmd <==
---
title: "Analyze dataset"
---

# Load packages

```{r load-packages}
#| label: load-packages
#| message: false

# Load packages
library(readr)        # reading/ writing datasets
library(stringr)      # match words  
library(knitr)        # for `kable()` function
library(dplyr)        # basic data manipulation
library(tidytext)     # text manipulation
library(qtalrkit)     # data dictionary creation
library(ggplot2)      # plotting
library(fs)           # file system operations
library(tibble)       # print data frames
library(kableExtra)   # creating formatted tables
library(infer)      # for statistical inference
library(skimr)      # for descriptive statistics
library(janitor)    # for cross-tabulation
```

# Read files

```{r}
#| label: read-files
#| message: false

meta_tbl <-
  read_csv("data/derived/meta_inf.csv")

meta_tbl <-
  meta_tbl |> 
  mutate(Category = factor(Category))|>
  mutate(Text_type = factor(Text_type))

meta_tbl 
```

# Identify

```{r}
#| label: map-hypothesis
#| message: false

# Dataset
meta_tbl 

# Specify the relationship
meta_spec <-
  meta_tbl |>
  specify(Num ~ Text_type)

# Preview
meta_spec
```

# Inspect

```{r}
#| label: assess-distribution
#| message: false

# Density plot
meta_tbl |>
  ggplot(aes(x = Num, fill = Text_type)) +
  geom_density(alpha = 0.75) +
  labs(
  x = "Metadiscourse markers number",
  y = "Density",
  fill = "Text type"
  )

meta_tbl |>
  ggplot(aes(x = Text_type, y = Num, fill = Text_type)) +
  geom_boxplot(notch = TRUE) +
  labs(
  x = "Text type",
  y = "Metadiscourse markers number",
  fill = "Text type"
  )

ggplot(meta_tbl, aes(x = Num, color = Text_type)) +
  stat_ecdf() +
  labs(
  x = "Metadiscourse markers number",
  y = "ECDF",
  color = "Text type"
  )
```

# Interrogate

## Observed statistic

```{r}
#| label: calculate-observed-statistic
#| message: false

meta_obs <-
  meta_spec |>
  # Target Text - Source Text
  calculate(stat = "diff in means", order = c("TT", "ST"))

meta_obs
```

## Null distribution

```{r}
#| label: generate-null-distribution
#| message: false

meta_null <-
  meta_spec |>
  hypothesize(null = "independence") |>
  generate(reps = 1000, type = "permute") |>
  calculate(stat = "diff in means", order = c("TT", "ST"))

meta_null |> visualize() +
  shade_p_value(obs_stat = meta_obs, direction = "greater")
```

## P-value

```{r}
#| label: calculate-p-value
#| message: false

meta_p <-
  meta_null |>
  get_p_value(obs_stat = meta_obs, direction = "greater")

meta_p
```

## Resampling distribution

```{r}
#| label: generate-resampling-distribution
#| message: false

meta_boot <-
  meta_spec |>
  generate(reps = 1000, type = "bootstrap") |>
  calculate(stat = "diff in means", order = c("TT", "ST"))

meta_boot |> visualize()
```

## Confidence interval

```{r}
#| label: calculate-confidence-interval
#| message: false

meta_ci <-
  meta_boot |>
  get_confidence_interval(level = 0.95)

meta_ci
```

# Interpret

```{r}
#| label: evaluate
#| message: false

# Evaluate the p-value
meta_p

# Evaluate the confidence interval
meta_ci

meta_boot |>
  visualize() +
  shade_confidence_interval(meta_ci)
```
